requirements.txt
fastapi[all]
asyncpg
tortoise-orm[asyncpg]